Split CLI diagnostics tests
